+++ CodeEdit/Features/Documents/Controllers/CodeEditSplitViewController.swift
@@ -52,23 +52,17 @@ final class CodeEditSplitViewController: NSSplitViewController {
               let editorManager = workspace.editorManager,
               let statusBarViewModel = workspace.statusBarViewModel,
               let utilityAreaModel = workspace.utilityAreaModel,
-              let taskmanager = workspace.taskManager else {
+              let taskManager = workspace.taskManager else {
             return
         }
 
         splitView.translatesAutoresizingMaskIntoConstraints = false
 
-        let settingsView = SettingsInjector {
+        let navigator = makeNavigator(view: SettingsInjector {
             NavigatorAreaView(workspace: workspace, viewModel: navigatorViewModel)
                 .environmentObject(workspace)
                 .environmentObject(editorManager)
-        }
-
-        let navigator = NSSplitViewItem(sidebarWithViewController: NSHostingController(rootView: settingsView))
-        navigator.titlebarSeparatorStyle = .none
-        navigator.isSpringLoaded = true
-        navigator.minimumThickness = Self.minSidebarWidth
-        navigator.collapseBehavior = .useConstraints
+        })
 
         addSplitViewItem(navigator)
 
@@ -89,20 +83,32 @@ final class CodeEditSplitViewController: NSSplitViewController {
 
         addSplitViewItem(mainContent)
 
-        let inspectorView = SettingsInjector {
+        let inspector = makeInspector(view: SettingsInjector {
             InspectorAreaView(viewModel: InspectorAreaViewModel())
                 .environmentObject(workspace)
                 .environmentObject(editorManager)
-        }
+        })
 
-        let inspector = NSSplitViewItem(inspectorWithViewController: NSHostingController(rootView: inspectorView))
+        addSplitViewItem(inspector)
+    }
+
+    private func makeNavigator(view: some View) -> NSSplitViewItem {
+        let navigator = NSSplitViewItem(sidebarWithViewController: NSHostingController(rootView: view))
+        navigator.titlebarSeparatorStyle = .none
+        navigator.isSpringLoaded = true
+        navigator.minimumThickness = Self.minSidebarWidth
+        navigator.collapseBehavior = .useConstraints
+        return navigator
+    }
+
+    private func makeInspector(view: some View) -> NSSplitViewItem {
+        let inspector = NSSplitViewItem(inspectorWithViewController: NSHostingController(rootView: view))
         inspector.titlebarSeparatorStyle = .none
         inspector.minimumThickness = Self.minSidebarWidth
         inspector.maximumThickness = .greatestFiniteMagnitude
         inspector.collapseBehavior = .useConstraints
         inspector.isSpringLoaded = true
-
-        addSplitViewItem(inspector)
+        return inspector
     }
 
     override func viewWillAppear() {
